requirements.txt
asyncio==3.4.3
Flask==3.0.0
Flask-Cors==4.0.0
langchain
langchain-openai
langchain_pinecone
langchain-anthropic
langchain-community
langchain_pinecone
langchainhub
langgraph
beautifulsoup4==4.12.3
jmespath==1.0.1
load_dotenv==0.1.0
lark==1.1.9
pinecone-client==3.2.2
pydash==7.0.7
pymongo
bcrypt